requirements.txt
scipy
numpy
ipython
pillow
